Fixed Sign Out Added new tests Move some tests from AuthenticationClientTest to SyncAuthenticationClientTest
