+++ library/src/test/java/com/okta/oidc/AuthenticateClientTest.java
@@ -184,31 +184,6 @@ public class AuthenticateClientTest {
     }
 
     @Test
-    public void refreshTokenRequest() throws InterruptedException, JSONException, AuthorizationException {
-        RefreshTokenRequest request = mSyncAuthClient.refreshTokenRequest();
-        String nonce = CodeVerifierUtil.generateRandomState();
-        String jws = TestValues.getJwt(mEndPoint.getUrl(), nonce, mAccount.getClientId());
-        mEndPoint.enqueueTokenSuccess(jws);
-        TokenResponse response = request.executeRequest();
-        assertNotNull(response);
-
-        TokenResponse original = mGson.fromJson(String.format(TOKEN_SUCCESS, jws),
-                TokenResponse.class);
-        assertEquals(original.getIdToken(), response.getIdToken());
-        assertEquals(original.getRefreshToken(), response.getRefreshToken());
-        assertEquals(original.getIdToken(), response.getIdToken());
-    }
-
-    @Test
-    public void refreshTokenRequestFailure() throws InterruptedException, JSONException, AuthorizationException {
-        mExpectedEx.expect(AuthorizationException.class);
-        mEndPoint.enqueueReturnInvalidClient();
-        RefreshTokenRequest request = mSyncAuthClient.refreshTokenRequest();
-        TokenResponse response = request.executeRequest();
-        assertNull(response);
-    }
-
-    @Test
     public void refreshToken() throws InterruptedException {
         final CountDownLatch latch = new CountDownLatch(1);
         String nonce = CodeVerifierUtil.generateRandomState();
@@ -235,183 +210,6 @@ public class AuthenticateClientTest {
         assertNull(cb.getResult());
         assertNotNull(cb.getException());
         assertEquals(cb.getException().getMessage(), "Network error");
-    }
-
-    @Test
-    public void configurationRequest() throws AuthorizationException, InterruptedException {
-        mEndPoint.enqueueConfigurationSuccess();
-        ConfigurationRequest request = mSyncAuthClient.configurationRequest();
-        ProviderConfiguration configuration = request.executeRequest();
-        RecordedRequest recordedRequest = mEndPoint.takeRequest();
-        assertThat(recordedRequest.getPath(),
-                equalTo("//.well-known/openid-configuration?client_id=CLIENT_ID"));
-        assertNotNull(configuration);
-        assertEquals(mGson.fromJson(PROVIDER_CONFIG, ProviderConfiguration.class).persist(),
-                configuration.persist());
-    }
-
-    @Test
-    public void configurationRequestFailure() throws AuthorizationException, InterruptedException {
-        mExpectedEx.expect(AuthorizationException.class);
-        mEndPoint.enqueueConfigurationFailure();
-        ConfigurationRequest request = mSyncAuthClient.configurationRequest();
-        ProviderConfiguration configuration = request.executeRequest();
-        RecordedRequest recordedRequest = mEndPoint.takeRequest();
-        assertThat(recordedRequest.getPath(),
-                equalTo("//.well-known/openid-configuration?client_id=CLIENT_ID"));
-        assertNull(configuration);
-    }
-
-    @Test
-    public void userProfileRequest() throws InterruptedException, AuthorizationException,
-            JSONException {
-        mEndPoint.enqueueUserInfoSuccess();
-        AuthorizedRequest request = mSyncAuthClient.userProfileRequest();
-        JSONObject result = request.executeRequest();
-        RecordedRequest recordedRequest = mEndPoint.takeRequest();
-        assertThat(recordedRequest.getHeader("Authorization"), is("Bearer " + ACCESS_TOKEN));
-        assertThat(recordedRequest.getHeader("Accept"), is(HttpConnection.JSON_CONTENT_TYPE));
-        assertThat(recordedRequest.getPath(), equalTo("/userinfo"));
-        assertNotNull(result);
-        assertEquals("John Doe", result.getString("name"));
-        assertEquals("Jimmy", result.getString("nickname"));
-    }
-
-    @Test
-    public void userProfileRequestFailure() throws InterruptedException, AuthorizationException {
-        mExpectedEx.expect(AuthorizationException.class);
-        mEndPoint.enqueueReturnUnauthorizedRevoked();
-        AuthorizedRequest request = mSyncAuthClient.userProfileRequest();
-        JSONObject result = request.executeRequest();
-        RecordedRequest recordedRequest = mEndPoint.takeRequest();
-        assertNull(result);
-        assertThat(recordedRequest.getPath(), equalTo("/userinfo"));
-    }
-
-    @Test
-    public void revokeTokenRequest() throws AuthorizationException, InterruptedException {
-        mEndPoint.enqueueReturnSuccessEmptyBody();
-        RevokeTokenRequest request = mSyncAuthClient.revokeTokenRequest("access_token");
-        boolean status = request.executeRequest();
-        RecordedRequest recordedRequest = mEndPoint.takeRequest();
-        assertThat(recordedRequest.getPath(),
-                equalTo("/revoke?client_id=CLIENT_ID&token=access_token"));
-        assertTrue(status);
-    }
-
-    @Test
-    public void revokeTokenRequestFailure() throws AuthorizationException, InterruptedException {
-        mExpectedEx.expect(AuthorizationException.class);
-        mEndPoint.enqueueReturnInvalidClient();
-        RevokeTokenRequest request = mSyncAuthClient.revokeTokenRequest("access_token");
-        boolean status = request.executeRequest();
-        RecordedRequest recordedRequest = mEndPoint.takeRequest();
-        assertFalse(status);
-        assertThat(recordedRequest.getPath(),
-                equalTo("/revoke?client_id=CLIENT_ID&token=access_token"));
-    }
-
-    @Test
-    public void authorizedRequest() throws InterruptedException, AuthorizationException,
-            JSONException {
-        //use userinfo for generic authorized request
-        mEndPoint.enqueueUserInfoSuccess();
-        Uri uri = Uri.parse(mProviderConfig.userinfo_endpoint);
-        HashMap<String, String> properties = new HashMap<>();
-        properties.put("state", CUSTOM_STATE);
-        AuthorizedRequest request = mSyncAuthClient.authorizedRequest(uri, properties,
-                null, HttpConnection.RequestMethod.GET);
-        JSONObject result = request.executeRequest();
-        RecordedRequest recordedRequest = mEndPoint.takeRequest();
-        assertThat(recordedRequest.getHeader("state"), is(CUSTOM_STATE));
-        assertThat(recordedRequest.getHeader("Authorization"), is("Bearer " + ACCESS_TOKEN));
-        assertThat(recordedRequest.getHeader("Accept"), is(HttpConnection.JSON_CONTENT_TYPE));
-        assertThat(recordedRequest.getPath(), equalTo("/userinfo"));
-        assertNotNull(result);
-        assertEquals("John Doe", result.getString("name"));
-        assertEquals("Jimmy", result.getString("nickname"));
-    }
-
-    @Test
-    public void authorizedRequestFailure() throws InterruptedException, AuthorizationException {
-        //use userinfo for generic authorized request
-        mExpectedEx.expect(AuthorizationException.class);
-        mEndPoint.enqueueReturnUnauthorizedRevoked();
-        Uri uri = Uri.parse(mProviderConfig.userinfo_endpoint);
-        HashMap<String, String> properties = new HashMap<>();
-        properties.put("state", CUSTOM_STATE);
-        AuthorizedRequest request = mSyncAuthClient.authorizedRequest(uri, properties,
-                null, HttpConnection.RequestMethod.GET);
-        JSONObject result = request.executeRequest();
-        RecordedRequest recordedRequest = mEndPoint.takeRequest();
-        assertThat(recordedRequest.getHeader("state"), is(CUSTOM_STATE));
-        assertThat(recordedRequest.getHeader("Authorization"), is("Bearer " + ACCESS_TOKEN));
-        assertThat(recordedRequest.getHeader("Accept"), is(HttpConnection.JSON_CONTENT_TYPE));
-        assertThat(recordedRequest.getPath(), equalTo("/userinfo"));
-        assertNull(result);
-    }
-
-    @Test
-    public void tokenExchangeFailure() throws InterruptedException, JSONException, AuthorizationException {
-        mExpectedEx.expect(AuthorizationException.class);
-        String codeVerifier = CodeVerifierUtil.generateRandomCodeVerifier();
-        String nonce = CodeVerifierUtil.generateRandomState();
-
-        AuthorizeRequest request = new AuthorizeRequest.Builder().codeVerifier(codeVerifier)
-                .authorizeEndpoint(mAccount.getDiscoveryUri().toString())
-                .redirectUri(mAccount.getRedirectUri().toString())
-                .scope(SCOPES)
-                .nonce(nonce)
-                .create();
-
-        AuthorizeResponse response = AuthorizeResponse.
-                fromUri(Uri.parse("com.okta.test:/callback?code=CODE&state=CUSTOM_STATE"));
-
-        mEndPoint.enqueueReturnInvalidClient();
-        TokenRequest tokenRequest = (TokenRequest) HttpRequestBuilder.newRequest()
-                .request(TOKEN_EXCHANGE).account(mAccount)
-                .providerConfiguration(mProviderConfig)
-                .authRequest(request)
-                .authResponse(response)
-                .createRequest();
-
-        TokenResponse tokenResponse = tokenRequest.executeRequest();
-        RecordedRequest recordedRequest = mEndPoint.takeRequest();
-        assertThat(recordedRequest.getPath(), equalTo("/token"));
-        assertNull(tokenResponse);
-    }
-
-    @Test
-    public void tokenExchangeSuccess() throws InterruptedException, JSONException, AuthorizationException {
-        String codeVerifier = CodeVerifierUtil.generateRandomCodeVerifier();
-        String nonce = CodeVerifierUtil.generateRandomState();
-
-
-        AuthorizeRequest request = new AuthorizeRequest.Builder().codeVerifier(codeVerifier)
-                .authorizeEndpoint(mAccount.getDiscoveryUri().toString())
-                .redirectUri(mAccount.getRedirectUri().toString())
-                .scope("openid", "email", "profile")
-                .nonce(nonce)
-                .create();
-
-        AuthorizeResponse response = AuthorizeResponse.
-                fromUri(Uri.parse("com.okta.test:/callback?code=CODE&state=CUSTOM_STATE"));
-
-        String jws = TestValues.getJwt(mEndPoint.getUrl(), nonce, mAccount.getClientId());
-
-        mEndPoint.enqueueTokenSuccess(jws);
-        TokenRequest tokenRequest = (TokenRequest) HttpRequestBuilder.newRequest()
-                .request(TOKEN_EXCHANGE).account(mAccount)
-                .authRequest(request)
-                .providerConfiguration(mProviderConfig)
-                .authResponse(response)
-                .createRequest();
-        TokenResponse tokenResponse = tokenRequest.executeRequest();
-
-        RecordedRequest recordedRequest = mEndPoint.takeRequest();
-        assertThat(recordedRequest.getPath(), equalTo("/token"));
-        assertNotNull(tokenResponse);
-        assertEquals(tokenResponse.getIdToken(), jws);
     }
 
     @Test
